Move ocp-build-data types and derefencing into api package
